changed a CreateConstantBuffer method signature to take a size_t instead, and fixed some compile errrors.

+++ GPUPro-6/src/Engine/Core/Graphics/Buffers/ConstantBuffer.cpp
@@ -12,11 +12,11 @@ void ConstantBuffer::ReleaseBuffer()
 }
 
 
-void ConstantBuffer::InitBuffer(UINT length)
+void ConstantBuffer::InitBuffer(size_t length)
 {
 	D3D11_BUFFER_DESC desc;
 	ZeroMemory(&desc, sizeof(D3D11_BUFFER_DESC));
-	desc.ByteWidth = length;
+	desc.ByteWidth = static_cast<UINT>(length);
 	desc.Usage = D3D11_USAGE_DYNAMIC;
 	desc.BindFlags = D3D11_BIND_CONSTANT_BUFFER;
 	desc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
